+++ src/components/player-statistics-card/player-statistics-card.component.js
@@ -1,4 +1,5 @@
-var moment = require("moment");
+var _ = require("underscore");
+var types = require("../scorecard/components/match_types");
 
 module.exports = {
     onCreate(input) {
@@ -7,6 +8,14 @@ module.exports = {
         var player = input.player;
 
         var isWinner = leg ? leg.winner_player_id === player.id : match.winner_id === player.id;
+
+        if (match.match_type.id === types.SHANGHAI) {
+            // If match finished with a shanghai, set the number on each player to display statistics correctly
+            var shanghai = _.find(input.statistics, (statistic) => statistic.shanghai !== null);
+            if (shanghai) {
+                _.each(input.statistics, (statistic) => statistic.shanghai = shanghai.shanghai );
+            }
+        }
         this.state = {
             isWinner: isWinner
         }