+++ src/pages/leg-result/components/leg-result/leg-result.component.js
@@ -79,7 +79,8 @@ module.exports = {
                 values[visit.player_id].push(current);
             }
             else {
-                if (match.match_type.id == types.SHOOTOUT || match.match_type.id == types.DARTS_AT_X || match.match_type.id === types.CRICKET) {
+                if (match.match_type.id == types.SHOOTOUT || match.match_type.id == types.DARTS_AT_X || match.match_type.id === types.CRICKET ||
+                    match.match_type.id == types.AROUND_THE_CLOCK || match.match_type.id == types.AROUND_THE_WORLD || match.match_type.id == types.SHANGHAI) {
                     current = current + visit.score;
                     values[visit.player_id].push(current);
                     if (current > chartMaxValue) {